WIP refactor: split some logic from Easy_EKS.ts into Easy_EKS_Cluster.ts and Easy_EKS_Essentials.ts
