+++ config/eks/higher_envs_eks_config.ts
@@ -62,7 +62,7 @@ export function deploy_essentials(config: Easy_EKS_Config_Data, stack: cdk.Stack
     const ALBC_Version = 'v2.12.0'; //April 9th, 2025 latest from https://github.com/kubernetes-sigs/aws-load-balancer-controller/releases
     const ALBC_IAM_Policy_Url = `https://raw.githubusercontent.com/kubernetes-sigs/aws-load-balancer-controller/refs/tags/${ALBC_Version}/docs/install/iam_policy.json`
     const ALBC_IAM_Policy_JSON = JSON.parse(request("GET", ALBC_IAM_Policy_Url).body.toString());
-    const ALBC_IAM_Policy = new iam.Policy(stack, `${config.id}_AWS_LB_Controller_policy_for_EKS`, {
+    const ALBC_IAM_Policy = new iam.Policy(stack, 'AWS_LB_Controller_IAM_policy_for_EKS', {
         document: iam.PolicyDocument.fromJson( ALBC_IAM_Policy_JSON ),
     });
     const ALBC_Kube_SA = new eks.ServiceAccount(stack, 'aws-load-balancer-controller', {
